+++ concordium-consensus/src/Concordium/GlobalState/ContractStateV1.hs
@@ -19,6 +19,7 @@ module Concordium.GlobalState.ContractStateV1
    freezeInMemoryPersistent,
    thawInMemoryPersistent,
    toByteString,
+   migratePersistentState,
    -- * Testing
    lookupKey,
    generatePersistentTree
@@ -81,6 +82,19 @@ newtype PersistentState = PersistentState (ForeignPtr PersistentState)
 -- in-memory.
 newtype InMemoryPersistentState = InMemoryPersistentState PersistentState
 
+-- |Migrate the provided persistent state from the existing backing store (which
+-- can be accessed using the provided 'LoadCallback'), to the new backing store
+-- (that is written to using the provided 'StoreCallback'). The input persistent
+-- state remains valid. The new persistent state is not cached, it is entirely
+-- stored on disk.
+foreign import ccall "migrate_persistent_tree_v1" migratePersistentTree :: LoadCallback -> StoreCallback -> Ptr PersistentState -> IO (Ptr PersistentState)
+
+migratePersistentState :: LoadCallback -> StoreCallback -> PersistentState -> IO PersistentState
+migratePersistentState lcbk scbk ps = do
+  newPSPtr <- withPersistentState ps $ migratePersistentTree lcbk scbk
+  newPS <- newForeignPtr freePersistentState newPSPtr
+  return (PersistentState newPS)
+
 -- |Gain temporary access to a pointer to the persistent state. The pointer
 -- should not be leaked from the computation.
 withPersistentState :: PersistentState -> (Ptr PersistentState -> IO a) -> IO a
@@ -113,10 +127,6 @@ foreign import ccall "thaw_persistent_state_v1" thawPersistentTree :: Ptr Persis
 -- |Get the amount of additional space that will be needed to store the new
 -- entries.
 foreign import ccall "get_new_state_size_v1" getNewStateSizeFFI :: LoadCallback -> Ptr MutableStateInner -> IO Word64
-
--- |Cache the persistent state, loading all parts of the tree that are purely on
--- disk.
-foreign import ccall "cache_persistent_state_v1" cachePersistentState :: LoadCallback -> Ptr PersistentState -> IO ()
 
 -- |Compute and retrieve the hash of the persistent state. The function is given
 -- a buffer to write the hash into.
@@ -182,11 +192,11 @@ instance (MonadBlobStore m) => BlobStorable m PersistentState where
       bRef <- withPersistentState ps $ storePersistentTree storeCallback
       return (put bRef, ps)
 
+-- |We are using the default no-op 'Cacheable' instance here
+-- as we do not want to load smart contract state
+-- into memory prematurely. The smart contract state is loaded
+-- on demand and flushed to disk upon finalization.
 instance MonadBlobStore m => Cacheable m PersistentState where
-  cache ps = do
-    (cbk, _) <- getCallbacks
-    liftIO (withPersistentState ps (cachePersistentState cbk))
-    return ps
 
 instance MonadBlobStore m => MHashableTo m SHA256.Hash PersistentState where
   getHashM ps = do