+++ concordium-consensus/src/Concordium/GlobalState/Paired.hs
@@ -1111,9 +1111,6 @@ instance (C.HasGlobalStateContext (PairGSContext lc rc) r,
     markPending pb = do
         coerceGSML $ markPending pb
         coerceGSMR $ markPending pb
-    clearAllNonFinalizedBlocks  = do
-        coerceGSML clearAllNonFinalizedBlocks
-        coerceGSMR clearAllNonFinalizedBlocks
     getGenesisBlockPointer = do
         gen1 <- coerceGSML getGenesisBlockPointer
         gen2 <- coerceGSMR getGenesisBlockPointer
@@ -1190,9 +1187,6 @@ instance (C.HasGlobalStateContext (PairGSContext lc rc) r,
             (Nothing, Nothing) -> return Nothing
             (Just pb1, Just pb2) -> assertEq pb1 pb2 $ return $ Just pb1
             _ -> error "takeNextPendingUntil (Paired): implementations returned different results"
-    wipePendingBlocks = do
-        coerceGSML wipePendingBlocks
-        coerceGSMR wipePendingBlocks
     getFocusBlock = do
         fb1 <- coerceGSML $ getFocusBlock
         fb2 <- coerceGSMR $ getFocusBlock
@@ -1251,18 +1245,23 @@ instance (C.HasGlobalStateContext (PairGSContext lc rc) r,
       coerceGSML (purgeTransactionTable t i)
       coerceGSMR (purgeTransactionTable t i)
 
-    wipeNonFinalizedTransactions = do
-        l1 <- coerceGSML wipeNonFinalizedTransactions
-        l2 <- coerceGSMR wipeNonFinalizedTransactions
-        -- Note that this test assumes the ordering is consistent
-        -- between implementations, which may not in general be a
-        -- reasonable assumption.
-        assertEq l1 l2 $ return l1
+    clearOnProtocolUpdate = do
+        coerceGSML clearOnProtocolUpdate
+        coerceGSMR clearOnProtocolUpdate
+
+    clearAfterProtocolUpdate = do
+        coerceGSML clearAfterProtocolUpdate
+        coerceGSMR clearAfterProtocolUpdate
 
     getNonFinalizedTransactionVerificationResult tx = do
       r1 <- coerceGSML $ getNonFinalizedTransactionVerificationResult tx
       r2 <- coerceGSMR $ getNonFinalizedTransactionVerificationResult tx
       assertEq r1 r2 $ return r1
+
+    storeFinalState (bs1, bs2) = do
+      coerceGSML (storeFinalState bs1)
+      coerceGSMR (storeFinalState bs2)
+
 newtype PairGSConfig c1 c2 = PairGSConfig (c1, c2)
 
 instance (GlobalStateConfig c1, GlobalStateConfig c2) => GlobalStateConfig (PairGSConfig c1 c2) where
@@ -1281,6 +1280,11 @@ instance (GlobalStateConfig c1, GlobalStateConfig c2) => GlobalStateConfig (Pair
         (ctx1, s1) <- initialiseNewGlobalState genData conf1
         (ctx2, s2) <- initialiseNewGlobalState genData conf2
         return (PairGSContext ctx1 ctx2, PairGState s1 s2)
+
+    migrateExistingState (PairGSConfig (conf1, conf2)) (PairGSContext ctx1 ctx2) (PairGState s1 s2) migrationData gd = do
+      (newCtx1, newState1) <- migrateExistingState conf1 ctx1 s1 migrationData gd
+      (newCtx2, newState2) <- migrateExistingState conf2 ctx2 s2 migrationData gd
+      return (PairGSContext newCtx1 newCtx2, PairGState newState1 newState2)
 
     activateGlobalState :: forall pv .
         IsProtocolVersion pv =>