+++ concordium-consensus/src/Concordium/GlobalState/Types.hs
@@ -3,6 +3,7 @@
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
+{-# LANGUAGE GADTs #-}
 
 module Concordium.GlobalState.Types where
 
@@ -15,6 +16,7 @@ import Concordium.GlobalState.BlockPointer (BlockPointerData)
 import Concordium.Wasm (WasmVersion)
 import Concordium.GlobalState.Classes
 import Concordium.Types
+import Concordium.Genesis.Data
 
 class (IsProtocolVersion (MPV m)) => MonadProtocolVersion (m :: Type -> Type) where
     type MPV m :: ProtocolVersion
@@ -37,6 +39,19 @@ class BlockStateTypes (m :: Type -> Type) where
     type BakerInfoRef m :: Type
     -- |A reference to an instrumented module.
     type InstrumentedModuleRef m :: WasmVersion -> Type
+
+-- |Data retrieved from the existing skov instance that is needed to construct
+-- the new instance. This is an existential type that closes over the **new**
+-- protocol version. The existing protocol version is determined from the monad
+-- @m@, which will have to be an instance of 'MonadProtocolVersion'.
+data PVInit m = forall pv . (IsProtocolVersion pv) => PVInit {
+    -- |Genesis data for the new chain.
+    pvInitGenesis :: Regenesis pv,
+    -- |Instructions on how to migrate from the existing instance to the new one.
+    pvInitMigration :: StateMigrationParameters (MPV m) pv,
+    -- |(Relative) height of the last finalized block.
+    pvInitFinalHeight :: BlockHeight
+  }
 
 -- |Account together with its index in the account map.
 type IndexedAccount m = (AccountIndex, Account m)