+++ concordium-consensus/tests/consensus/ConcordiumTests/ReceiveTransactionsTest.hs
@@ -220,7 +220,7 @@ runMyMonad act time initialState = runReaderT (runDeterministic (runStateT (runP
 
 -- |Run the given computation in a state consisting of only the genesis block and the state determined by it.
 runMyMonad' :: MyMonad a -> UTCTime -> GenesisData PV -> IO (a, MyState)
-runMyMonad' act time gd = runPureBlockStateMonad (initialSkovDataDefault gd (hashBlockState $ mock bs) genTT) >>= runMyMonad act time
+runMyMonad' act time gd = runPureBlockStateMonad (initialSkovDataDefault (genesisConfiguration gd) (hashBlockState $ mock bs) genTT Nothing) >>= runMyMonad act time
   where
     (bs, genTT) = case genesisState gd of
                Left err -> error $ "Invalid genesis state: " ++ err