+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- Speed up and reduce memory overhead during protocol updates.
+
 - Smart contract modules are no longer retained in memory. Module artifacts are loaded as needed
   during contract execution. Metadata is cached for a limited number of smart contract modules.
   By default, the cache will retain metadata for at most 100 smart contract modules, and this is
@@ -21,6 +23,7 @@
   `--chunksize` option specifies the size of export files in blocks.
 - The `--download-blocks-from` option now takes the URL to the catchup _index file_, permitting to
   only download and import catchup files containing blocks not already present in the database.
+- Smart contract state is no longer cached on startup and is not cached after finalization.
 
 ## 4.3.1